+++ lib/src/call.dart
@@ -1,8 +1,8 @@
 // Dart imports:
 import 'dart:async';
 import 'dart:core';
-import 'dart:ui';
 import 'dart:io' show Platform;
+import 'dart:ui';
 
 // Flutter imports:
 import 'package:flutter/cupertino.dart';
@@ -30,6 +30,9 @@ import 'package:zego_uikit_prebuilt_call/src/invitation/service.dart';
 import 'package:zego_uikit_prebuilt_call/src/minimizing/data.dart';
 import 'package:zego_uikit_prebuilt_call/src/minimizing/defines.dart';
 import 'package:zego_uikit_prebuilt_call/src/minimizing/overlay_machine.dart';
+
+import 'controller/private/pip/pip_android.dart';
+import 'controller/private/pip/pip_ios.dart';
 
 /// Call Widget.
 /// You can embed this widget into any page of your project to integrate the functionality of a call.
@@ -112,6 +115,8 @@ class ZegoUIKitPrebuiltCall extends StatefulWidget {
 
 class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
     with SingleTickerProviderStateMixin {
+  var contextInitNotifier = ValueNotifier<bool>(false);
+
   var barVisibilityNotifier = ValueNotifier<bool>(true);
   var barRestartHideTimerNotifier = ValueNotifier<int>(0);
   var chatViewVisibleNotifier = ValueNotifier<bool>(false);
@@ -150,13 +155,24 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
     return isAllEntered;
   }
 
+  bool get playingStreamInPIPUnderIOS {
+    bool isPlaying = false;
+    if (Platform.isIOS) {
+      isPlaying = (ZegoUIKitPrebuiltCallController().pip.private.pipImpl()
+              as ZegoCallControllerIOSPIP)
+          .isSupportInConfig;
+    }
+
+    return isPlaying;
+  }
+
   @override
   void initState() {
     super.initState();
 
     ZegoUIKit().getZegoUIKitVersion().then((version) {
       ZegoLoggerService.logInfo(
-        'version: zego_uikit_prebuilt_call:4.16.5; $version, \n'
+        'version: zego_uikit_prebuilt_call:4.17.0; $version, \n'
         'config:${widget.config}, \n'
         'events:${widget.events}, \n',
         tag: 'call',
@@ -205,6 +221,7 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
         subTag: 'prebuilt',
       );
 
+      contextInitNotifier.value = true;
       listenUserEvents();
     } else {
       controller.private.initByPrebuilt(
@@ -219,6 +236,7 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
       );
       controller.minimize.private.initByPrebuilt(
         minimizeData: minimizeData,
+        config: widget.config,
       );
       controller.permission.private.initByPrebuilt(
         config: widget.config,
@@ -229,7 +247,21 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
 
       /// not wake from mini page
       initContext().then((_) {
+        ZegoLoggerService.logInfo(
+          'initContext done',
+          tag: 'call',
+          subTag: 'prebuilt',
+        );
+
+        contextInitNotifier.value = true;
+
         listenUserEvents();
+      }).catchError((e) {
+        ZegoLoggerService.logError(
+          'initContext exception:$e',
+          tag: 'call',
+          subTag: 'prebuilt',
+        );
       });
     }
     ZegoCallMiniOverlayMachine().changeState(ZegoCallMiniOverlayPageState.idle);
@@ -278,7 +310,12 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
       controller.permission.private.uninitByPrebuilt();
       controller.pip.private.uninitByPrebuilt();
 
-      ZegoUIKit().leaveRoom();
+      ZegoUIKit().leaveRoom().then((_) {
+        if (playingStreamInPIPUnderIOS) {
+          ZegoUIKit().enableHardwareDecoder(false);
+          ZegoUIKit().enableCustomVideoRender(false);
+        }
+      });
       // await ZegoUIKit().uninit();
     } else {
       ZegoLoggerService.logInfo(
@@ -347,17 +384,28 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
 
   @override
   Widget build(BuildContext context) {
-    if (Platform.isAndroid) {
-      return PiPSwitcher(
-        floating: ZegoUIKitPrebuiltCallController().pip.private.floating,
-        childWhenDisabled: normalPage(),
-        childWhenEnabled: screenUtil(
-          childWidget: pipPage(),
-        ),
-      );
-    }
+    return ValueListenableBuilder<bool>(
+      valueListenable: contextInitNotifier,
+      builder: (context, isDone, _) {
+        if (isDone) {
+          if (Platform.isAndroid) {
+            return PiPSwitcher(
+              floating: (ZegoUIKitPrebuiltCallController().pip.private.pipImpl()
+                      as ZegoCallControllerPIPAndroid)
+                  .floating,
+              childWhenDisabled: normalPage(),
+              childWhenEnabled: screenUtil(
+                childWidget: pipPage(),
+              ),
+            );
+          }
 
-    return normalPage();
+          return normalPage();
+        }
+
+        return const Center(child: CircularProgressIndicator());
+      },
+    );
   }
 
   Widget screenUtil({required Widget childWidget}) {
@@ -489,10 +537,20 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
       await ZegoUIKit().setAdvanceConfigs(widget.config.advanceConfigs);
 
       ZegoUIKit()
-          .init(appID: widget.appID, appSign: widget.appSign)
+          .init(
+        appID: widget.appID,
+        appSign: widget.appSign,
+        enablePlatformView: playingStreamInPIPUnderIOS,
+        playingStreamInPIPUnderIOS: playingStreamInPIPUnderIOS,
+      )
           .then((value) async {
         /// second set after create express
         await ZegoUIKit().setAdvanceConfigs(widget.config.advanceConfigs);
+
+        if (playingStreamInPIPUnderIOS) {
+          await ZegoUIKit().enableHardwareDecoder(true);
+          await ZegoUIKit().enableCustomVideoRender(true);
+        }
 
         _setVideoConfig();
         _setBeautyConfig();
@@ -764,6 +822,8 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
       );
     }
 
+    final _defaultAudioVideoContainer = defaultAudioVideoContainer();
+
     return Positioned.fromRect(
       rect: widget.config.audioVideoView.containerRect?.call() ??
           Rect.fromLTWH(0, 0, preferWidth, preferHeight),
@@ -780,7 +840,7 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
                     ZegoUIKit().getAudioVideoList(),
                     audioVideoViewCreator,
                   ) ??
-                  defaultAudioVideoContainer();
+                  _defaultAudioVideoContainer;
             },
           );
         },