+++ lib/src/components/bottom_menu_bar.dart
@@ -19,11 +19,11 @@ import 'package:zego_uikit_prebuilt_call/src/controller.dart';
 import 'package:zego_uikit_prebuilt_call/src/defines.dart';
 import 'package:zego_uikit_prebuilt_call/src/events.dart';
 import 'package:zego_uikit_prebuilt_call/src/events.defines.dart';
+import 'package:zego_uikit_prebuilt_call/src/invitation/service.dart';
 import 'package:zego_uikit_prebuilt_call/src/minimizing/data.dart';
 import 'package:zego_uikit_prebuilt_call/src/minimizing/defines.dart';
 import 'package:zego_uikit_prebuilt_call/src/minimizing/mini_button.dart';
 import 'package:zego_uikit_prebuilt_call/src/minimizing/overlay_machine.dart';
-import 'package:zego_uikit_prebuilt_call/src/invitation/service.dart';
 
 /// @nodoc
 class ZegoCallBottomMenuBar extends StatefulWidget {
@@ -135,7 +135,11 @@ class _ZegoCallBottomMenuBarState extends State<ZegoCallBottomMenuBar> {
   List<Widget> getDisplayButtons(BuildContext context) {
     final needRestoreDeviceState =
         widget.minimizeData.isPrebuiltFromMinimizing ||
-            ZegoUIKitPrebuiltCallController().pip.private.isRestoreFromPIP;
+            ZegoUIKitPrebuiltCallController()
+                .pip
+                .private
+                .pipImpl()
+                .isRestoredFromPIP;
 
     final buttonList = <Widget>[
       ...getDefaultButtons(