+++ lib/src/invitation/components/send_call_invitation_button.dart
@@ -291,6 +291,7 @@ class _ZegoSendCallInvitationButtonState
         ZegoUIKit().getSignalingPlugin().getConnectionState()) {
       ZegoLoggerService.logError(
         'signaling is not connected:${ZegoUIKit().getSignalingPlugin().getConnectionState()}, '
+        'ZegoUIKitPrebuiltCallInvitationService is init: ${ZegoUIKitPrebuiltCallInvitationService().isInit}'
         'please call ZegoUIKitPrebuiltCallInvitationService.init with ZegoUIKitSignalingPlugin first',
         tag: 'call-invitation',
         subTag: 'components, send call button',